Reorganized loggers
