+++ loggers/elasticsearch/honeypy_elasticsearch.py
