+++ loggers/honeydb/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2016 foospidy
+# https://github.com/foospidy/HoneyPy
+# See LICENSE for details
+
+import sys
+# prevent creation of compiled bytecode files
+sys.dont_write_bytecode = True
