+++ loggers/logstash/honeypy_logstash.py
